requirements.txt
# tensorflow==2.10.0
# numpy==1.24.4
# pandas>=1.3.0
# scikit-learn>=1.0.0
# matplotlib>=3.4.0
# h5py>=2.9.0
# uvicorn[standard]
# fastapi
# tensorflow==2.10.0
tensorflow-cpu==2.10.0

numpy==1.24.4
pandas==1.5.3
scikit-learn==1.2.2
matplotlib==3.7.1
h5py==3.8.0
uvicorn[standard]==0.22.0
fastapi==0.95.2
